There was a bug in the scenario calculation. To figure out alignment it needs to divide the sum of the indicators by the sum of the responses. However, some of the responses weren't there in the comb_scenario dataframe. That's because they weren't needed in the personal alignment calculation.
I commented out the line that did the filtering, and then I had to add a couple of lines that would handle the missing values (i.e. make them 0s).

This seems to work. It may not be the most efficient solution though as it substantially increases the size of the dataframes.